Commit2

+++ http/view_permissions.php
@@ -20,30 +20,27 @@ if (checkAuthorization(9, 'view permissions')) {
 				pg_free_result(pgquery("INSERT INTO table_user(tablename, username) VALUES('{$_GET['tablename']}', '{$_GET['username']}');"));
 				echo 'Row (\'', htmlspecialchars($_GET['tablename']), '\', \'', htmlspecialchars($_GET['username']), "') inserted.<br/>\n";
 			}
-			pg_free_result($result);
-		} else if (!empty($_GET['key1']) && !empty($_GET['key2'])) {
+		} else if (!empty($_GET['key1']) && !empty($_GET['key2']) && isset($_GET['update'])) {
 			$result = pgquery("SELECT TRUE FROM users WHERE username = '{$_GET['key2']}' OR username = '{$_GET['username']}' AND is_administrator;");
 			if ($_GET['key2'] == $_SESSION['username'] && $_GET['username'] == $_SESSION['username'] || $_SESSION['is_administrator'] && !pg_fetch_row($result) || $_SESSION['is_root']) {
-				if (isset($_GET['update'])) {
-					pg_free_result(pgquery("UPDATE table_user SET (tablename, username) = ('{$_GET['tablename']}', '{$_GET['username']}') WHERE tablename = '{$_GET['key1']}' AND username = '{$_GET['key2']}';"));
-					echo 'Row (\'', htmlspecialchars($_GET['key1']), '\', \'', htmlspecialchars($_GET['key2']), "') updated.<br/>\n";
-				} else if (isset($_GET['delete'])) {
-					if (isset($_GET['confirm'])) {
-						pg_free_result(pgquery("DELETE FROM table_user WHERE tablename = '{$_GET['key1']}' AND username = '{$_GET['key2']}';"));
-						echo 'Row (\'', htmlspecialchars($_GET['key1']), '\', \'', htmlspecialchars($_GET['key2']), "') deleted.<br/>\n";
-					} else {
-?>
-						Are you sure?
-<?php
-						echo '<a href="?key1=', urlencode($_GET['key1']), '&amp;key2=', urlencode($_GET['key2']), "&amp;delete&amp;confirm\">Yes</a>\n";
-?>
-						<a href="">No</a>
-<?php
-						exit(0);
-					}
-				}
+				pg_free_result(pgquery("UPDATE table_user SET (tablename, username) = ('{$_GET['tablename']}', '{$_GET['username']}') WHERE tablename = '{$_GET['key1']}' AND username = '{$_GET['key2']}';"));
+				echo 'Row (\'', htmlspecialchars($_GET['key1']), '\', \'', htmlspecialchars($_GET['key2']), "') updated.<br/>\n";
 			}
-			pg_free_result($result);
+		}
+		pg_free_result($result);
+	} else if (!empty($_GET['key1']) && !empty($_GET['key2']) && isset($_GET['delete'])) {
+		if (isset($_GET['confirm'])) {
+			pg_free_result(pgquery("DELETE FROM table_user WHERE tablename = '{$_GET['key1']}' AND username = '{$_GET['key2']}';"));
+			echo 'Row (\'', htmlspecialchars($_GET['key1']), '\', \'', htmlspecialchars($_GET['key2']), "') deleted.<br/>\n";
+		} else {
+?>
+			Are you sure?
+<?php
+			echo '<a href="?key1=', urlencode($_GET['key1']), '&amp;key2=', urlencode($_GET['key2']), "&amp;delete&amp;confirm\">Yes</a>\n";
+?>
+			<a href="">No</a>
+<?php
+			exit(0);
 		}
 	}
 	if ($_SESSION['is_root']) {
